+++ YACReader/render.cpp
@@ -861,12 +861,12 @@ unsigned int Render::numPages()
     return comic->numPages();
 }
 
-QByteArray Render::getCurrentRawPage() const
+QByteArray Render::getRawPage(int page) const
 {
     if (comic == nullptr)
         return QByteArray();
 
-    return comic->getRawPage(currentIndex);
+    return comic->getRawPage(page);
 }
 
 bool Render::hasLoadedComic()